pyproject.toml
[project]
name = "quranmedialib"
version = "0.1.0"
description = "Media producing library for Quranic texts."
readme = "README.md"
requires-python = ">=3.13, <3.14"
dependencies = [
    "Pillow<13.0.0",
]
